gui-shop/requirements.txt
tk
fpdf
qrcode
Pillow
python-dotenv
